A template version
This is a template version of my allocator. You can use it in your
project freely. The allocator can allocate any single class’s object,
with flexible capacity and I’m working on auto deletion of the objects
when the allocator itself is deleted. Note that the allocator’s memory
uses will keep expand to the largest usage of your allocated objects
using it. You can manually renew the allocator to release memory but
note when you do this, all pointers allocated by it expires.